+++ frontend/src/providers/SidebarProvider.tsx
@@ -1,0 +1,13 @@
+import { useState, ReactNode } from 'react';
+import SidebarContext from '../contexts/SidebarContext';
+
+export const SidebarProvider = ({ children }: { children: ReactNode }) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const toggle = () => setIsOpen(!isOpen);
+
+  return (
+    <SidebarContext.Provider value={{ isOpen, toggle }}>
+      {children}
+    </SidebarContext.Provider>
+  );
+};
